Extract respacer from format.ts
